requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-core
pydantic